upgrade

+++ src/admin/api/objects/master.php
@@ -11,6 +11,16 @@ class Master {
         $this->conn = $db;
     }
 
+    function exists() {
+        $query = "SELECT * FROM " . $this->table_name . " WHERE id = {$this->id}";
+        $stmt = $this->conn->prepare($query);
+        $stmt->execute();
+        if ($stmt->fetch()) {
+            return true;
+        }
+        return false;;
+    }
+
     function read() {
         $query = "SELECT * FROM " . $this->table_name;
         $stmt = $this->conn->prepare($query);
@@ -34,6 +44,9 @@ class Master {
     }
 
     function update() {
+        if (! exists()) {
+            return false;
+        }
         $query = "UPDATE " . $this->table_name . "
         SET
             name=\"{$this->name}\",
@@ -52,6 +65,9 @@ class Master {
     }
 
     function delete() {
+        if (! $this->exists()) {
+            return false;
+        }
         $query = "DELETE FROM " . $this->table_name . " WHERE id = {$this->id}";
         $stmt = $this->conn->prepare($query);
         $this->id = htmlspecialchars(strip_tags($this->id));
