+++ src/admin/api/objects/order.php
@@ -15,6 +15,16 @@ class Order {
         $this->conn = $db;
     }
 
+    function exists() {
+        $query = "SELECT * FROM " . $this->table_name . " WHERE id = {$this->id}";
+        $stmt = $this->conn->prepare($query);
+        $stmt->execute();
+        if ($stmt->fetch()) {
+            return true;
+        }
+        return false;;
+    }
+
     function read() {
         $query = "SELECT * FROM " . $this->table_name;
         $stmt = $this->conn->prepare($query);
@@ -42,6 +52,9 @@ class Order {
     }
 
     function update() {
+        if (! $this->exists()) {
+            return false;
+        }
         $query = "UPDATE " . $this->table_name . "
         SET
             info=\"{$this->info}\",
@@ -68,6 +81,9 @@ class Order {
     }
 
     function delete() {
+        if (! $this->exists()) {
+            return false;
+        }
         $query = "DELETE FROM " . $this->table_name . " WHERE id = {$this->id}";
         $stmt = $this->conn->prepare($query);
         $this->id = htmlspecialchars(strip_tags($this->id));
